+++ src/cncView/FileContentTree.ts
@@ -12,66 +12,64 @@ const parser = require(('./ncParser'));
 export class FileContentProvider implements vscode.TreeDataProvider<vscode.TreeItem>{
     private _onDidChangeTreeData: vscode.EventEmitter<vscode.TreeItem | undefined | null | void> = new vscode.EventEmitter<vscode.TreeItem | undefined | null | void>();
     readonly onDidChangeTreeData: vscode.Event<vscode.TreeItem | undefined | null | void> = this._onDidChangeTreeData.event;
-    fileItem: FileItem;
+    fileItem: FileItem = new FileItem("", vscode.TreeItemCollapsibleState.None);
     matchCategories: MatchCategories;
     context: vscode.ExtensionContext;
     currentFileWatcher: fs.FSWatcher | undefined;
     file: vscode.Uri | undefined;
-    constructor(file: vscode.Uri | undefined, extContext: vscode.ExtensionContext) {
-        this.file = file;
+    constructor(extContext: vscode.ExtensionContext) {
         this.matchCategories = {
             toolCalls: new CategoryItem("Tool Calls"),
             prgCalls: new CategoryItem("Program Calls"),
         };
         this.context = extContext;
-        this.fileItem = this.createFileItem();
+        this.update();
         this.updateFileWatcher();
-        this.updateTreeView(this.file);
     }
 
-    private createFileItem(): FileItem {
-        let fileItem;
+    private async createFileItem(): Promise<void> {
         if (this.file === undefined) {
-            fileItem = new FileItem("There is no currently opened file", this.matchCategories, vscode.TreeItemCollapsibleState.None);
+            this.fileItem = new FileItem("There is no currently opened file", vscode.TreeItemCollapsibleState.None);
         } else if (!isNcFile(this.file.fsPath)) {
-            fileItem = new FileItem("The currently opened file is no NC-file", this.matchCategories, vscode.TreeItemCollapsibleState.None);
+            this.fileItem = new FileItem("The currently opened file is no NC-file", vscode.TreeItemCollapsibleState.None);
         } else {
-            fileItem = new FileItem(Path.basename(this.file.fsPath), this.matchCategories, vscode.TreeItemCollapsibleState.Expanded);
+            await new Promise(r => setTimeout(r, 50)); //to prevent reading in between "file cleared" and "new content saved"
+            const filecontent = fs.readFileSync(this.file.fsPath, "utf8");
+            let parsed: boolean = true;
+            let parseResult;
+            try {
+                parseResult = parser.parse(filecontent);
+            } catch (error: any) {
+                if (error instanceof parser.SyntaxError) {
+                    this.fileItem = new FileItem("The currently opened NC-file has wrong syntax", vscode.TreeItemCollapsibleState.None);
+                    parsed = false;
+                }
+            }
+            if (parsed) {
+                this.updateMatchItems(parseResult);
+                this.fileItem = new FileItem(Path.basename(this.file.fsPath), vscode.TreeItemCollapsibleState.Expanded, this.matchCategories);
+            }
         }
-        return fileItem;
     }
 
     private updateFileWatcher() {
         if (this.file !== undefined) {
             this.currentFileWatcher?.close();
             this.currentFileWatcher = fs.watch(this.file.fsPath, () => {
-                this.updateTreeView(this.file);
+                this.update();
             });
         }
     }
 
     /**
      * Updates the tree so it shows the information concerning the specified file
-     * @param file 
      */
-    async updateTreeView(file: vscode.Uri | undefined) {
+    async update() {
         try {
-            this.file = file;
+            await new Promise(r => setTimeout(r, 10));
+            this.file =  vscode.window.activeTextEditor?.document.uri;;
             this.disposeCommands();
-            this.fileItem = this.createFileItem();
-            if (this.file !== undefined) {
-                await new Promise(r => setTimeout(r, 50)); //to prevent reading in between "file cleared" and "new content saved"
-                const filecontent = fs.readFileSync(this.file.fsPath, "utf8");
-                try {
-                    let parseResult = parser.parse(filecontent);
-
-                    this.updateMatchItems(parseResult);
-                } catch (error: any) {
-                    if (error instanceof parser.SyntaxError) {
-                        this.fileItem = new FileItem("The currently opened NC-file has wrong syntax", this.matchCategories, vscode.TreeItemCollapsibleState.None);
-                    }
-                }
-            }
+            this.createFileItem();
             this.updateFileWatcher();
             this._onDidChangeTreeData.fire();  //triggers updating the graphic
         } catch (error: any) {
@@ -115,12 +113,14 @@ export class FileContentProvider implements vscode.TreeDataProvider<vscode.TreeI
  */
 class FileItem extends vscode.TreeItem implements MyItem {
     private _children: Array<MyItem>;
-    constructor(label: string, matchCategories: MatchCategories, collapsibleState: vscode.TreeItemCollapsibleState) {
+    constructor(label: string, collapsibleState: vscode.TreeItemCollapsibleState, matchCategories?: MatchCategories) {
         super(label, collapsibleState);
         this._children = new Array<MyItem>();
-        Object.entries(matchCategories).forEach(([key, category]) => {
-            this.addChild(category);
-        });
+        if (matchCategories !== undefined) {
+            Object.entries(matchCategories).forEach(([key, category]) => {
+                this.addChild(category);
+            });
+        }
     }
 
     private addChild(category: CategoryItem): void {
@@ -444,5 +444,5 @@ function getLine(file: string, lineNumber: number): string {
  * @returns true if given uri ends with.nc, false otherwise
  */
 function isNcFile(path: string): boolean {
-    return Path.extname(path) === ".nc";
+    return Path.extname(path.toLowerCase()) === ".nc";
 }