+++ src/extension.ts
@@ -44,8 +44,9 @@ const nonAsciiCharacterDecorationType = vscode.window.createTextEditorDecoration
 let selectedLinesStatusBarItem: vscode.StatusBarItem;
 let currentOffsetStatusBarItem: vscode.StatusBarItem;
 
+//NC-file sidebar tree provider
 let fileContentProvider: FileContentProvider;
-let fileContentTreeView: vscode.TreeView<vscode.TreeItem> | undefined;
+let fileContentTreeView: vscode.TreeView<vscode.TreeItem>;
 
 // package.json information
 let packageFile;
@@ -105,8 +106,10 @@ export function activate(context: vscode.ExtensionContext): void {
 
     //NC-file sidebar tree provider
     const currentFile = vscode.window.activeTextEditor?.document.uri;
-    fileContentProvider = new FileContentProvider(currentFile, extContext);
-
+    fileContentProvider = new FileContentProvider(extContext);
+    fileContentTreeView = vscode.window.createTreeView('cnc-show-filecontent', {
+        treeDataProvider: fileContentProvider
+    });
     // commands
     context.subscriptions.push(
         vscode.commands.registerCommand("isg-cnc.FindAllToolCalls", () =>
@@ -307,11 +310,7 @@ function updateConfig() {
 function activeTextEditorChanged() {
     //Tree view
     try {
-        const currentFile = vscode.window.activeTextEditor?.document.uri;
-        fileContentProvider.updateTreeView(currentFile);
-        fileContentTreeView = vscode.window.createTreeView('cnc-show-filecontent', {
-            treeDataProvider: fileContentProvider
-        });
+        fileContentProvider.update();
     } catch (e: any) {
         vscode.window.showErrorMessage(e);
     }