+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Types of changes
 
 ### Added
 
+- Autogenerate github pages at push to main branch
+- Increase indent at linebreak and decrease indent if linebreak finished
+
 ### Fixed
 
 - Add toLowerCase for nc file detection